backend/requirements-sprint0.txt
# 🚀 Sprint-0 Requirements - BIST AI Smart Trader
# Veri katmanı, WebSocket, Fundamental data ve Grey TOPSIS için gerekli paketler

# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
asyncio
aiohttp>=3.8.0
websockets>=10.0
python-dotenv>=0.19.0

# Data processing & ML
scikit-learn>=1.1.0
scipy>=1.9.0

# Financial data & analysis
yfinance>=0.1.87
financetoolkit>=1.0.0

# MCDM (Multi-Criteria Decision Making)
pymcdm>=1.0.0

# WebSocket & async
websockets>=10.0
aiohttp>=3.8.0

# Data visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Database & storage
firebase-admin>=6.0.0
google-cloud-firestore>=2.0.0

# API framework
fastapi>=0.95.0
uvicorn>=0.20.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Logging & monitoring
loguru>=0.6.0
sentry-sdk>=1.20.0

# Environment & config
python-dotenv>=0.19.0
pydantic>=1.10.0

# Async utilities
asyncio-mqtt>=0.11.0
aiofiles>=0.8.0

# Data validation
marshmallow>=3.19.0
cerberus>=1.3.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.60.0

# Security
cryptography>=3.4.0
passlib>=1.7.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
